configure integration test
