+++ spring-exceptions/src/test/java/org/baeldung/ex/nontransientdataaccessexception/DataSourceLookupExceptionIntegrationTest.java
@@ -13,7 +13,7 @@ import org.springframework.test.context.support.AnnotationConfigContextLoader;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = { Cause4NonTransientConfig.class }, loader = AnnotationConfigContextLoader.class)
-public class DataSourceLookupExceptionTest {
+public class DataSourceLookupExceptionIntegrationTest {
 
     @Test(expected = DataSourceLookupFailureException.class)
     public void whenLookupNonExistentDataSource_thenDataSourceLookupFailureException() {
